+++ src/beat_link_trigger/players.clj
@@ -481,16 +481,19 @@
         (.drawString g frac-frame (int 122) (int 40))))))
 
 (defn- tempo-values
-  "Look up the current playback pitch percentage and effective tempo
-  for the specified player, returning them when available as a vector
-  followed by boolean indications of whether the device is the
-  current tempo master and whether it is synced."
+  "Look up the current playback pitch percentage and effective tempo for
+  the specified player, returning them when available as a vector
+  followed by boolean indications of whether the device is the current
+  tempo master and whether it is synced, and if that sync has degraded
+  to BPM-only."
   [^Long n]
   (when-let [^DeviceUpdate u (when (.isRunning time-finder) (.getLatestUpdateFor time-finder n))]
-    [(org.deepsymmetry.beatlink.Util/pitchToPercentage (.getPitch u))
-     (when (not= 65535 (.getBpm u)) (.getEffectiveTempo u))  ; Detect when tempo is not valid
-     (.isTempoMaster u)
-     (true? (when-let [cdj-status (.getLatestStatusFor virtual-cdj n)] (.isSynced cdj-status)))]))
+    (let [cdj-status (.getLatestStatusFor virtual-cdj n)]
+      [(org.deepsymmetry.beatlink.Util/pitchToPercentage (.getPitch u))
+       (when (not= 65535 (.getBpm u)) (.getEffectiveTempo u))  ; Detect when tempo is not valid
+       (.isTempoMaster u)
+       (true? (when cdj-status (.isSynced cdj-status)))
+       (true? (when cdj-status (.isBpmOnlySynced cdj-status)))])))
 
 (defn- format-pitch
   "Formats a number for display as a pitch value, with one or two
@@ -514,7 +517,7 @@
   the component being drawn, and the graphics context in which drawing
   is taking place."
   [n _ ^Graphics2D g]
-  (when-let [[pitch tempo master synced] (tempo-values n)]
+  (when-let [[pitch tempo master synced bpm-only] (tempo-values n)]
     (let [abs-pitch       (Math/abs (double pitch))]
       (.setRenderingHint g RenderingHints/KEY_ANTIALIASING RenderingHints/VALUE_ANTIALIAS_ON)
       (.setPaint g (Color/WHITE))
@@ -527,11 +530,12 @@
       (.setFont g (util/get-display-font :teko Font/PLAIN 14))
       (.drawString g "%" (int 56) (int 40)))
     (when synced
+      (when bpm-only (.setPaint g Color/YELLOW))
       (let [frame (java.awt.geom.RoundRectangle2D$Double. 40.0 4.0 24.0 15.0 4.0 4.0)]
         (.fill g frame)
         (.setPaint g Color/BLACK)
-        (.setFont g (util/get-display-font :teko Font/PLAIN 14))
-        (.drawString g "SYNC" (int 42) (int 16))))
+        (.setFont g (util/get-display-font :teko Font/PLAIN (if bpm-only 15 14)))
+        (.drawString g (if bpm-only "BPM" "SYNC") (int 42) (int 16))))
     (.setPaint g (if master Color/ORANGE Color/WHITE))
     (let [frame        (java.awt.geom.RoundRectangle2D$Double. 68.0 1.0 50.0 38.0 8.0 8.0)
           clip         (.getClip g)